requirements.txt
scikit-image
spacy
pandas
pytest
Pillow
requests
torch
setuptools
pycocotools
sauth
numpy
min_dalle

# App
streamlit

# Models and datasets
git+https://github.com/openai/CLIP.git